+++ HISTORY.rst
@@ -5,7 +5,7 @@ History
 0.5.0 (not yet)
 ---------------
 
-* blackwhite subcommand to convert color images to black and white
+* blackwhite subcommand to convert color images to black and white incl. doc
 
 0.4.0 (2019-02-26)
 ------------------